Improve CLI to have independent names
