+++ src/accelerate/commands/config/__init__.py
@@ -15,73 +15,42 @@
 # limitations under the License.
 
 import argparse
-import os
 
-from accelerate.utils import ComputeEnvironment
-
-from .cluster import get_cluster_input
-from .config_args import cache_dir, default_config_file, default_yaml_config_file, load_config_from_file  # noqa: F401
-from .config_utils import _ask_field, _ask_options, _convert_compute_environment  # noqa: F401
-from .sagemaker import get_sagemaker_input
+from .config import config_command, config_command_parser
+from .config_args import default_config_file, load_config_from_file  # noqa: F401
+from .default import default_command_parser, default_config_command
 
 
-description = "Launches a series of prompts to create and save a `default_config.yaml` configuration file for your training system. Should always be ran first on your machine"
+def filter_command_args(args: dict, args_prefix: str):
+    "Filters args while only keeping ones that are prefixed with `{args_prefix}.`"
+    new_args = argparse.Namespace()
+    for key, value in vars(args).items():
+        print(key, value)
+        if key.startswith(args_prefix):
+            setattr(new_args, key.replace(f"{args_prefix}.", ""), value)
+    return new_args
 
 
-def get_user_input():
-    compute_environment = _ask_options(
-        "In which compute environment are you running?",
-        ["This machine", "AWS (Amazon SageMaker)"],
-        _convert_compute_environment,
-    )
-    if compute_environment == ComputeEnvironment.AMAZON_SAGEMAKER:
-        config = get_sagemaker_input()
-    else:
-        config = get_cluster_input()
-    return config
+def get_config_parser(subparsers=None):
+    parent_parser = argparse.ArgumentParser(add_help=False)
+    # The main config parser
+    config_parser = config_command_parser(subparsers)
 
+    # Then add other parsers with the parent parser
+    default_parser = default_command_parser(config_parser, parents=[parent_parser])  # noqa: F841
 
-def config_command_parser(subparsers=None):
-    if subparsers is not None:
-        parser = subparsers.add_parser("config", description=description)
-    else:
-        parser = argparse.ArgumentParser("Accelerate config command", description=description)
-
-    parser.add_argument(
-        "--config_file",
-        default=None,
-        help=(
-            "The path to use to store the config file. Will default to a file named default_config.yaml in the cache "
-            "location, which is the content of the environment `HF_HOME` suffixed with 'accelerate', or if you don't have "
-            "such an environment variable, your cache directory ('~/.cache' or the content of `XDG_CACHE_HOME`) suffixed "
-            "with 'huggingface'."
-        ),
-    )
-
-    if subparsers is not None:
-        parser.set_defaults(func=config_command)
-    return parser
-
-
-def config_command(args):
-    config = get_user_input()
-    if args.config_file is not None:
-        config_file = args.config_file
-    else:
-        if not os.path.isdir(cache_dir):
-            os.makedirs(cache_dir)
-        config_file = default_yaml_config_file
-
-    if config_file.endswith(".json"):
-        config.to_json_file(config_file)
-    else:
-        config.to_yaml_file(config_file)
+    return config_parser
 
 
 def main():
-    parser = config_command_parser()
-    args = parser.parse_args()
-    config_command(args)
+    config_parser = get_config_parser()
+    args = config_parser.parse_args()
+    if not args.default:
+        args = filter_command_args(args, "config_args")
+        config_command(args)
+    elif args.default:
+        args = filter_command_args(args, "default_args")
+        default_config_command(args)
 
 
 if __name__ == "__main__":
